Fix uploading blob image path to server #64
When form submission is requested, user's selected profile image should
be sent to the server and converted into server image path before
requesting a PATCH request. Without this step, the server will recieve
a blob image that is maintained locally on the browser and won't display
on our deployment server.